requirements.txt
torch
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
flask>=3.0.0
Pillow>=10.0.0
